Voice Texting
Voice Texting is the newest feature of Conver. It adds the ability to message a contact hands-free and even when your phone isn't close by! It uses the awesome functionality of [Pushbullet](https://www.pushbullet.com/), a service built by a team of superheroes in San Francisco that allows SMS messaging and text management from a computer!